actually rename auth views
